FEAT : killcount, hitcount, victorycount등 게임 플레이시 저장하거나 확인해야할 수치들을 모두 StatisticsManager에게 전가시키고 저장 및 불러오기를 StatisticsManager의 정보들로 수행하게 변경, 통계를 보여주기 위한 panel 구현 (메인씬에서 열람가능)

+++ Assets/Scripts/Core/Observer/SubjectKill.cs
@@ -5,8 +5,8 @@ namespace Core.Observer
 {
     public class SubjectKill : ASubject
     {
-        private int _count = 0;
-        public int Count
+        private long _count = 0;
+        public long Count
         {
             get => _count;
             set
@@ -24,8 +24,8 @@ namespace Core.Observer
             GoalCount = goalCount;
         }
         
-        public int Count { get; set; }
-        public int GoalCount;
+        public long Count { get; set; }
+        public long GoalCount;
         public Action Action;
         
         // subject에게 알림을 받을때 사용하는 함수
